requirements.txt
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
iniconfig==1.0.1
instabot==0.117.0
isort==4.3.21
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
mock==4.0.2
more-itertools==8.4.0
moviepy==1.0.3
numpy==1.19.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
praw==7.1.0
prawcore==1.5.0
proglog==0.1.9
py==1.9.0
pycodestyle==2.6.0
pylint==2.5.3
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.0.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.10.16
schedule==0.6.0
six==1.15.0
toml==0.10.1
tqdm==4.48.2
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
wrapt==1.12.1
